% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-IntervalList.R
\docType{class}
\name{IntervalList-class}
\alias{IntervalList-class}
\title{S4 class representing a list of intervals}
\description{
S4 class representing a list of intervals
}
\section{Slots}{

\describe{
\item{\code{mid}:}{A vector of real numbers saved as a \code{numeric}
object specifying the mid-points of the intervals.}
\item{\code{spr}:}{A vector of real numbers saved as a \code{numeric}
object specifying the spreads of the intervals.}
}
}

\examples{
showClass("IntervalList")
showMethods(classes = "IntervalList")
}
\seealso{
Objects of \code{IntervalList} class should be created through
\code{\link{IntervalList}()} function. Besides \code{IntervalList} class,
the \pkg{IntervalQuestionStat} package also incorporates
\code{\link{IntervalData-class}} and \code{\link{IntervalMatrix-class}}
for dealing with interval-valued data in R environment.
}
\author{
José García-García \email{garciagarjose@uniovi.es}
}
