% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-IntervalMatrix.R
\docType{class}
\name{IntervalMatrix-class}
\alias{IntervalMatrix-class}
\title{S4 class representing a matrix of intervals}
\description{
S4 class representing a matrix of intervals
}
\section{Slots}{

\describe{
\item{\code{mid}:}{A matrix of real numbers saved as a \code{matrix}
object specifying the mid-points of the intervals.}
\item{\code{spr}:}{A matrix of real numbers saved as a \code{matrix}
object specifying the spreads of the intervals.}
}
}

\examples{
showClass("IntervalMatrix")
showMethods(classes = "IntervalMatrix")
}
\seealso{
Objects of \code{IntervalMatrix} class should be created through
\code{\link{IntervalMatrix}()} function. Besides \code{IntervalMatrix} class,
the \pkg{IntervalQuestionStat} package also incorporates
\code{\link{IntervalData-class}} and \code{\link{IntervalList-class}}
for dealing with interval-valued data in R environment.
}
\author{
José García-García \email{garciagarjose@uniovi.es}
}
