% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-as.IntervalData.R
\docType{methods}
\name{as.IntervalData}
\alias{as.IntervalData}
\alias{as.IntervalData,numeric-method}
\title{Convert an object to \code{IntervalData} class}
\usage{
\S4method{as.IntervalData}{numeric}(object)
}
\arguments{
\item{object}{a single real number.}
}
\value{
Returns an object of class \code{\linkS4class{IntervalData}}.
}
\description{
This function coerce a real number to an object of \code{IntervalData} class.
}
\examples{
as.IntervalData(1)
}
\seealso{
Other IntervalData-method: 
\code{\link{IntervalData-class}},
\code{\link{IntervalData}},
\code{\link{arithmetic}},
\code{\link{distance}()},
\code{\link{ivd2likert}()},
\code{\link{ivd2vas}()},
\code{\link{plot}()},
\code{\link{show}()}

Other conversion: 
\code{\link{as.IntervalList}()}
}
\author{
Jose Garcia Garcia \email{garciagarjose@uniovi.es}
}
\concept{IntervalData-method}
\concept{conversion}
