% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-as.IntervalList.R
\docType{methods}
\name{as.IntervalList}
\alias{as.IntervalList}
\alias{as.IntervalList,IntervalData-method}
\title{Convert an object to \code{IntervalList} class}
\usage{
\S4method{as.IntervalList}{IntervalData}(object)
}
\arguments{
\item{object}{an interval-valued data.}
}
\value{
Returns an object of class \code{\linkS4class{IntervalList}}.
}
\description{
This function coerce an object of \code{IntervalData} to an object of \code{IntervalList} class.
}
\examples{
as.IntervalList(IntervalData(0, 1))
}
\seealso{
Other IntervalList-method: 
\code{\link{IntervalList-class}},
\code{\link{IntervalList}},
\code{\link{cov}()},
\code{\link{ivd2likert}()},
\code{\link{ivd2vas}()},
\code{\link{mean}()},
\code{\link{plot}()},
\code{\link{sum}()},
\code{\link{var}()}

Other conversion: 
\code{\link{as.IntervalData}()}
}
\author{
Jose Garcia Garcia \email{garciagarjose@uniovi.es}
}
\concept{IntervalList-method}
\concept{conversion}
