% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-IntervalList.R
\docType{class}
\name{IntervalList-class}
\alias{IntervalList-class}
\title{S4 class representing a list of interval-valued data}
\description{
S4 class representing a list of interval-valued data
}
\section{Slots}{

\describe{
\item{\code{.Data}:}{The data part, which is an object of class 'list'.}}
}

\examples{
showClass("IntervalList")
showMethods(classes = "IntervalList")
}
\seealso{
Other IntervalList-method: 
\code{\link{IntervalList}},
\code{\link{as.IntervalList}()},
\code{\link{cov}()},
\code{\link{ivd2likert}()},
\code{\link{ivd2vas}()},
\code{\link{mean}()},
\code{\link{plot}()},
\code{\link{sum}()},
\code{\link{var}()}
}
\author{
Jose Garcia Garcia \email{garciagarjose@uniovi.es}
}
\concept{IntervalList-method}
