% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-IntervalMatrix.R
\docType{class}
\name{IntervalMatrix-class}
\alias{IntervalMatrix-class}
\title{S4 class representing a matrix of interval-valued data}
\description{
S4 class representing a matrix of interval-valued data
}
\section{Slots}{

\describe{
\item{\code{.Data}:}{The data part, which is an object of class 'matrix'.}}
}

\examples{
showClass("IntervalMatrix")
showMethods(classes = "IntervalMatrix")
}
\seealso{
Other IntervalMatrix-method: 
\code{\link{IntervalMatrix}},
\code{\link{apply}()},
\code{\link{ivd2likert}()},
\code{\link{ivd2vas}()},
\code{\link{show}()}
}
\author{
Jose Garcia Garcia \email{garciagarjose@uniovi.es}
}
\concept{IntervalMatrix-method}
